Began work on City addon unit!
City Addon Unit will contain things for building your city and expanding
it on a server, which don't directly relate to roads. Initially, I'm
focusing on "Things you'll find in a shopping center", so I'm starting
with what I believe to be Forge's first digital/physical hybrid economy
system. It works like real life- you have a bank account with money, and
you can withdraw the money into physical coins and notes. If I get
really clever, you'll need a card with a pin to withdraw money and
anyone with the pin can do it!