Bounding boxes for diagonal roads now render when hovered if empty
In other words, you can now see (and break) a diagonal block if there's nothing that would make that block render normally, but only when you aim at it.

Open to feedback; I can make it so this same wireframe permanently appears, but I'm not sure thats a nice solution.