Update source code MC/Forge mappings
Updated a few method mappings in the source code to
follow those given by the new toolchain. This is
primarily the result of the build scripts hard
referencing the stable-39 mappings for Forge 1.12.

Previously, the snapshot-20171003-1.12 mappings were
used which are more in-line with mappings from game
versions prior to 1.12. This slightly modernizes
the code base as a result of moving to stable-39.